Run the script on start and watch instead of install
